+++ CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Ability to extract the release notes of a version from a changelog, and fallback to "unreleased" if no version is specified.
+- Option to fallback to "unreleased" if the version specified was not found. 
 
 [unreleased]: https://github.com/bittacklr/peek-at-changelog/compare/ORIGIN...HEAD